RFCT: debug_trace_print.
