+++ circuits/src/generation/mod.rs
@@ -27,13 +27,13 @@ use self::tape::generate_tape_trace;
 
 pub mod builtin;
 pub mod cpu;
-mod debug_trace_print;
 pub mod memory;
 pub mod poseidon;
 pub mod poseidon_chunk;
 pub mod sccall;
 pub mod storage;
 pub mod tape;
+mod ctl_test;
 
 #[derive(Clone, Debug, Deserialize, Serialize, Default)]
 /// Inputs needed for trace generation.
